Updated IndexerSettingsModel with better descriptions.
Enable checkbox for Indexer is now labeled as such.
Removed jscript load for deleted script.
A user will now be unable to add a rootDir if the SabDropDir value is the same (ignoring case).
A user will also be unable to set their SabDropDir to the same value as an exiting RootDir (Ignoring case).
The last two changes prevent users from possibly causing all TV shows to be deleted.