Added locale language results when adding new serie
When searching for a show (add series) and if the locale language of the
system is not english, the same search will be done with the locale
language.

Results are filtered so only one language get returned, priority to
locale.

This will not make that metadata become non-english, since the serie
info is update shortly after the insertion with skyhook information, but
the LocaleCleanTitle will be saved.