New: "Force" priority for NZBGet
The force priority allows NZBs to be downloaded even if the general queue is in a paused state. The use case here is setting recently released items to be downloaded even if the queue is paused. This is a quite common pattern on ISPs that have download quotas (e.g.: most in Australia) where downloading huge archives may need to be paced but shows being currently aired should be fetched for viewing as soon as they are available.

The existence of this priority is buried in their API docs[1] and in forum posts[2] but has been supported for a while (as of r1000). On older clients, the NZBGet UI will just show "Priority: 900" instead of the friendly label. They will be prioritised above other items in the queue but won't bypass the paused state. In other words, NZBGet models priorities as numeric values so this is compatible with old versions and new versions will treat this value specially when encountered.

[1]: http://nzbget.net/RPC_API_reference#Method_.22append.22
[2]: http://nzbget.net/forum/viewtopic.php?f=3&t=1177