Mono: Try copying files with cp --reflink first.
This should save a bunch of space for people on compatible filesystems
(like Btrfs). `cp` with `--reflink=auto` will also just do a normal
copy of the file if reflinks aren't possible, so the fallback to the
base is more of a "just in case" than anything. A short (10s) timeout
is used if the source or destination mounts are network mounts, to avoid
cp hanging with network mounts where the server has gone down.