Allow torrent removal if finished.
Currently, torrents in Transmission need to meet both seeding time AND ratio requirements in order to be considered for deletion.  This presents a problem, because Sonarr relies on Transmission's own internal completion mechanism to stop torrents once the seeding goals are met. This means that if a torrent was seeded for enough time, but didn't meet the ratio, it would be stopped and marked as "finished" in Transmission, but Sonarr would never remove it since the ratio was not met (and never will be due to Transmission stopping it).