Change qBittorrent V1/V2 Selector to use cache
Use Cache object to store last detected qBittorrent API version. Currently set cache timeout to 10 minutes as a moderate compromise between possible updates to qBittorrent (Though it would have to be an update from < v4 to > 4.2 to break the functionality for the 10 minutes) and frquency of requests to qBittorrent to check the API version for each call.