Add name field to release profiles
This adds a new, non-functional field to release profiles for a friendly
name. This name may be used, for example, by scripts invoking the API to
look for specifically named profiles in order to find their ID, which is
required to perform further operations on it. Additionally it adds an
additional organization tool for users to manage their release profiles.

Signed-off-by: Robert Dailey <rcdailey@gmail.com>