fix: Throw ArgumentException if spec impl cannot be found
Previously, if the spec was not found, you'd get a
NullReferenceException back in the API response because the code always
assumed the `SingleOrDefault()` method would return non-null.

The use case for this is custom format import via API (and might also
apply to import on the frontend too, although I did not test this). If
you take a CF meant for Radarr and import it into Sonarr, you get this
issue.