Ensure CustomFormat is taken from disk filename
As it stands, sonarr only relies on the scene filename, which may be incorrect (i.e. mismatch between release title and release filename). This can be worked around, by having a custom renaming scheme. However, sonarr doesn't look at the filename on disk as a priority, and thus the CF score is always based on the original release filename. This PR aims to always use what's on disk to determine the imported CF score.