Add DevContainer and VSCode config
This makes it possible to develop Sonarr in a DevContainer with VSCode.

On an M1/M2 Mac, I found that the default linux-x64 runtime identifier
is not valid for the aarch64 docker image that will be used by default.
So instead of hardcoding the runtime identifier, I've added a check to
set the architecture based on the current process architecture.