fix: improve bitrate fetching for media files
Added a check for tags that start with "BPS" to retrieve the first occurrence of a bitrate value. This check uses the FirstOrDefault method to find tags where the key starts with "BPS", considering it may vary for different languages (e.g., BPS-jpn). While I have encountered only BPS-eng so far, this approach is flexible enough to handle other variations in the future.

However, if this method is deemed unstable, it could be hardcoded to specifically match BPS-eng instead.

Why this update might be helpful:
Fixes missing bitrate information on the Media Info popover in the web UI, for both audio and video streams. I have found that the BPS-eng tag is relatively common on anime specifically, which is consequently often missing bitrate information.