fix(connection): use ISO 8601 format specifier for Discord timestamps
This fixes the embed timestamp having an incorrect format in cultures
that do not use colons as time specifiers, such as da-DK. Colons (:)
have a special meaning when used within DateTime.ToString format
strings, being converted to the time delimiter of the CurrentCulture.
This is an issue, as Discord rejects embeds with an incorrect timestamp
with the cryptic `{"embeds": ["0"]}` error.